Create Todo App
At this stage I've created a basic todo app with features below:
Fetch and Display list of available Todos
Add a new todo item
Checkbox to select todo items in order to perform an action on them. P.s: This is not completed yet, but can be improved on

P.S: A real database was not used. I used json-server to model a database at this stage